# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrackMovil — API",
    "version": "0.1.0",
    "description": "Catálogo de las APIs propias de TrackMovil (RiogasTracking, app 5 en SecuritySuite).\n\nGENERADO por `pnpm docs:api` a partir de app/api/**/route.ts. No editar a mano:\nlas descripciones, los consumidores y los ejemplos van en docs/api/anotaciones.yaml,\nque se mergea encima al servir GET /api/docs/spec.\n\n`x-auth` refleja los gates que el handler REALMENTE invoca. `x-auth.sinGate: true`\nsignifica que el endpoint no valida nada: es información sensible y por eso este\nportal es solo-root."
  },
  "servers": [
    {
      "url": "https://track.glp.riogas.com.uy",
      "description": "producción"
    }
  ],
...
  "paths": {
...
    "/api/admin/login-security/usuario-detalle": {
      "get": {
        "operationId": "getApiAdminLoginSecurityUsuarioDetalle",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Request inválido"
          },
          "401": {
            "description": "No autenticado"
          },
          "502": {
            "description": "Error del upstream"
          }
        },
        "summary": "GET /api/admin/login-security/usuario-detalle",
        "tags": [
          "admin"
        ],
        "x-archivo": "app/api/admin/login-security/usuario-detalle/route.ts",
        "x-auth": {
          "funcionalidades": [
            "Query Inicios de sesion"
          ],
          "gates": [
            "requireAuthorizationHeader (token para el SecuritySuite)",
            "requireFuncionalidad"
          ],
          "sinGate": false
        },
        "x-generado": true
      }
    },
...
    "/api/admin/usuarios-empresa": {
      "get": {
        "operationId": "getApiAdminUsuariosEmpresa",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Request inválido"
          },
          "401": {
            "description": "No autenticado"
          },
          "500": {
            "description": "Error interno"
          },
          "502": {
            "description": "Error del upstream"
          }
        },
        "summary": "GET /api/admin/usuarios-empresa",
        "tags": [
          "admin"
        ],
        "x-archivo": "app/api/admin/usuarios-empresa/route.ts",
        "x-auth": {
          "funcionalidades": [
            "Gestion de Usuarios"
          ],
          "gates": [
            "header x-track-isroot (spoofeable)",
            "requireAuthorizationHeader (token para el SecuritySuite)",
            "requireFuncionalidad"
          ],
          "sinGate": false
        },
        "x-generado": true
      }
    },
    "/api/admin/usuarios-empresa/toggle": {
      "post": {
        "operationId": "postApiAdminUsuariosEmpresaToggle",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Request inválido"
          },
          "401": {
            "description": "No autenticado"
          },
          "502": {
            "description": "Error del upstream"
          }
        },
        "summary": "POST /api/admin/usuarios-empresa/toggle",
        "tags": [
          "admin"
        ],
        "x-archivo": "app/api/admin/usuarios-empresa/toggle/route.ts",
        "x-auth": {
          "funcionalidades": [
            "Gestion de Usuarios"
          ],
          "gates": [
            "requireAuthorizationHeader (token para el SecuritySuite)",
            "requireFuncionalidad"
          ],
          "sinGate": false
        },
        "x-generado": true
      }
    },
...
    "/api/empresas": {
      "get": {
        "operationId": "getApiEmpresas",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Error interno"
          }
        },
        "summary": "GET /api/empresas",
        "tags": [
          "empresas"
        ],
        "x-archivo": "app/api/empresas/route.ts",
        "x-auth": {
          "funcionalidades": [],
          "gates": [
            "requireAuth (sesión Supabase)"
          ],
          "sinGate": false
        },
        "x-generado": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sesionSupabase": {
        "type": "http",
        "scheme": "bearer",
        "description": "Sesión de Supabase Auth validada por requireAuth (lib/auth-middleware.ts)."
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key de integraciones externas (sender GeneXus/SGM, VB6)."
      },
      "tokenSecapi": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT emitido por SecuritySuite. Es el que verifica lib/docs/root-guard.ts contra secapi."
      }
    }
  },
...
}